Fix weinreb recipe + test
Also moved some deprecated preprocessing methods from _simple to _deprecated
